refactor + fix : write to big query
